move code to projects.ts
